build: fix `catalyst-core` dependency resolving
Previously, it was pointing to an `flatDir` with an absolute path, which causes inconsistencies between dependency resolving across many operating systems. The fix involved adding a dot before the path, which makes it comply with how relative paths work on Unix-like systems.

For example, using `/dependencies` without a `.` prefix would point to a directory called "dependencies" at the root of your filesystem, whereas prefixing the path with a dot would use the current directory as base.